changes in conjunction with CL work
