+++ contribs/bicycle/src/main/java/org/matsim/contrib/bicycle/BicycleModule.java
@@ -32,7 +32,7 @@ import org.matsim.vehicles.VehicleType;
  * @author smetzler, dziemke
  */
 final class BicycleModule extends AbstractModule {
-	// necessary to have this public
+	// needs to be public since otherwise nobody can overwrite parts of Bicycles.addAsOverridingModules(...).  kai, sep'19
 	
 	private static final Logger LOG = Logger.getLogger(BicycleModule.class);
 
@@ -49,8 +49,7 @@ final class BicycleModule extends AbstractModule {
 		addTravelDisutilityFactoryBinding(bicycleConfigGroup.getBicycleMode()).to(BicycleTravelDisutilityFactory.class).in(Singleton.class);
 		bindScoringFunctionFactory().to(BicycleScoringFunctionFactory.class).in(Singleton.class);
 
-		bind(BicycleLinkSpeedCalculator.class);
-		// yyyy not sure if this is helpful since if it is not bound against an interface, one cannot replace it anyways.  kai, sep'19
+		bind( BicycleLinkSpeedCalculator.class ).to( BicycleLinkSpeedCalculatorDefaultImpl.class ) ;
 
         if (bicycleConfigGroup.isMotorizedInteraction()) {
 			addMobsimListenerBinding().to(MotorizedInteractionEngine.class);
