+++ contribs/bicycle/src/main/java/org/matsim/contrib/bicycle/BicycleQSimModule.java
@@ -8,7 +8,9 @@ import org.matsim.core.mobsim.qsim.AbstractQSimModule;
 import org.matsim.core.mobsim.qsim.qnetsimengine.ConfigurableQNetworkFactory;
 import org.matsim.core.mobsim.qsim.qnetsimengine.QNetworkFactory;
 
-class BicycleQSimModule extends AbstractQSimModule {
+public class BicycleQSimModule extends AbstractQSimModule {
+	// needs to be public since otherwise nobody can overwrite parts of Bicycles.addAsOverridingModules(...).  kai, sep'19
+
     @Override
     protected void configureQSim() {
         // if used together with BicyclesModule the BicycleLinkSpeedCalculator is already bound.
@@ -25,7 +27,7 @@ class BicycleQSimModule extends AbstractQSimModule {
         private EventsManager events;
 
         @Inject
-        private BicycleLinkSpeedCalculator calculator;
+        private BicycleLinkSpeedCalculatorDefaultImpl calculator;
 
         @Override
         public QNetworkFactory get() {
